// app/core/hexag/registros/sociedades/domain/enums/society-register-step.enum.ts
export enum SocietyRegisterStep {
  DATOS_SOCIEDAD = "datos-sociedad",
  ACCIONISTAS = "accionistas",
  ACCIONES = "acciones",
  ASIGNACION_ACCIONES = "asignacion-acciones",
  DIRECTORIO = "directorio",
  REGISTRO_APODERADOS = "registro-apoderados",
  REGIMEN_PODERES = "regimen-poderes",
  QUORUMS_MAYORIAS = "quorums-mayorias",
  ACUERDOS_SOCIETARIOS = "acuerdos-societarios",
  RESUMEN = "resumen",
  FINALIZAR = "finalizar",
}